MusicThrouple/requirements.txt
Flask==2.1.2
Flask-Bcrypt==1.0.1
NumPy==1.24.0
pandas==1.4.2
psycopg2==2.9.3
python-dotenv==0.20.0
